Refactor: 닉네임 검증 로직 Validator 클래스로 별도 분리 (#36)
- 닉네임 검증을 Service 레이어에서 하는 게 맞다고 생각은 들지만, 그 검증 코드들을 포함하기엔 부담이 된다고 생각했습니다.
- 결론적으로 닉네임 검증을 위한 Validator 클래스를 만들어 별도로 분리했습니다.

issue #36
